+++ openedx_events/content_authoring/data.py
@@ -12,7 +12,12 @@ from typing import BinaryIO, List
 
 import attr
 from opaque_keys.edx.keys import CourseKey, UsageKey
-from opaque_keys.edx.locator import LibraryLocatorV2, LibraryUsageLocatorV2
+from opaque_keys.edx.locator import (
+    LibraryCollectionLocator,
+    LibraryContainerLocator,
+    LibraryLocatorV2,
+    LibraryUsageLocatorV2,
+)
 
 
 @attr.s(frozen=True)
@@ -160,8 +165,9 @@ class ContentLibraryData:
     Data related to a content library that has changed.
 
     Attributes:
-        library_key (LibraryLocatorV2): a key that represents a Blockstore-based content library.
-        update_blocks (bool): flag that indicates whether the content library blocks indexes should be updated.
+        library_key (LibraryLocatorV2): a key that represents a v2 content library.
+        update_blocks (bool): DEPRECATED flag to indicate whether the content library blocks indexes should be updated.
+            Now we send individual events for each updated item instead.
     """
 
     library_key = attr.ib(type=LibraryLocatorV2)
@@ -174,7 +180,7 @@ class LibraryBlockData:
     Data related to a library block that has changed.
 
     Attributes:
-        library_key (LibraryLocatorV2): a key that represents a Blockstore-based content library.
+        library_key (LibraryLocatorV2): a key that represents a v2 content library.
         usage_key (LibraryUsageLocatorV2): a key that represents a XBlock in a Blockstore-based content library.
     """
 
@@ -218,12 +224,27 @@ class LibraryCollectionData:
     Data related to a library collection that has changed.
 
     Attributes:
-        library_key (LibraryLocatorV2): a key that represents a Blockstore-based content library.
-        collection_key (str): identifies the collection within the library's learning package
-        background (bool): indicate whether the sender doesn't want to wait for handler to finish execution,
-           i.e., the handler can run the task in background. By default it is False.
+        collection_key (LibraryCollectionLocator): identifies the collection within the library's learning package
+        background (bool): **DEPRECATED** field. Indicated the sender doesn't want to wait for handler to finish
+           execution. Now instead we recommend that simple handlers are synchronous and the _sender_ of the event should
+           send the event(s) from an async celery task if it is expected to result in a lot of handlers being called.
     """
 
-    library_key = attr.ib(type=LibraryLocatorV2)
-    collection_key = attr.ib(type=str)
+    collection_key = attr.ib(type=LibraryCollectionLocator)
+    background = attr.ib(type=bool, default=False)
+
+
+@attr.s(frozen=True)
+class LibraryContainerData:
+    """
+    Data related to a library container that has changed.
+
+    Attributes:
+        container_key (LibraryContainerLocator): identifies the container (e.g.  unit, section)
+        background (bool): **DEPRECATED** field. Indicated the sender doesn't want to wait for handler to finish
+           execution. Now instead we recommend that simple handlers are synchronous and the _sender_ of the event should
+           send the event(s) from an async celery task if it is expected to result in a lot of handlers being called.
+    """
+
+    container_key = attr.ib(type=LibraryContainerLocator)
     background = attr.ib(type=bool, default=False)